Fix tar header error reporting on 32 bit systems
If an extension header is rejected because its too big, the error path
would print the size as size_t, altough it is an uint64_t. On 64 bit
systems, this works because size_t is a 64 bit unsigned integer, on
32 bit systems, not so much.

+++ lib/tar/read_header.c
@@ -448,16 +448,16 @@ out_eof:
 	clear_header(out);
 	return 1;
 fail_slink_len:
-	fprintf(stderr, "rejecting GNU symlink header with size %zu\n",
-		pax_size);
+	fprintf(stderr, "rejecting GNU symlink header with size %lu\n",
+		(unsigned long)pax_size);
 	goto fail;
 fail_path_len:
-	fprintf(stderr, "rejecting GNU long path header with size %zu\n",
-		pax_size);
+	fprintf(stderr, "rejecting GNU long path header with size %lu\n",
+		(unsigned long)pax_size);
 	goto fail;
 fail_pax_len:
-	fprintf(stderr, "rejecting PAX header with size %zu\n",
-		pax_size);
+	fprintf(stderr, "rejecting PAX header with size %lu\n",
+		(unsigned long)pax_size);
 	goto fail;
 fail_magic:
 	fputs("input is not a ustar tar archive!\n", stderr);
